requirements.txt
openai==1.55.1
supabase==1.2.0
python-dotenv==1.0.1
pgvector==0.2.5
pyyaml==6.0.1
pytest==8.0.1
python-jose==3.3.0
loguru==0.7.2
pydantic==2.4.2
pydantic-core==2.10.1
httpx==0.24.1
postgrest==0.11.0
